Split commands and contracts between packages
