+++ src/main.py
@@ -5,24 +5,24 @@ from importlib.metadata import metadata
 import click
 
 import src
-from src.commands.consolidate import consolidate
 from src.commands.create_keys import create_keys
 from src.commands.create_wallet import create_wallet
-from src.commands.exit_validators import exit_validators
 from src.commands.init import init
-from src.commands.process_meta_vaults import process_meta_vaults
 from src.commands.recover import recover
-from src.commands.setup_remote_signer import setup_remote_signer
 from src.commands.start.hashi_vault import start_hashi_vault
 from src.commands.start.local import start_local
 from src.commands.start.relayer import start_relayer
 from src.commands.start.remote_signer import start_remote_signer
 from src.commands.submit_rated_network import submit_rated_network
 from src.common.utils import get_build_version
+from src.meta_vault.commands.process_meta_vaults import process_meta_vaults
 from src.nodewise.commands.node_install import node_install
 from src.nodewise.commands.node_start import node_start
 from src.nodewise.commands.node_status import node_status_command as node_status
 from src.remote_db.commands import remote_db_group
+from src.validators.commands.consolidate import consolidate
+from src.validators.commands.exit_validators import exit_validators
+from src.validators.commands.setup_remote_signer import setup_remote_signer
 
 build = get_build_version()
 version = src.__version__
