+++ src/meta_vault/commands/process_meta_vaults.py
