+++ src/validators/commands/consolidate.py
